[FIX] + [TEST]
- Trigger convention for queries + Query logging
- Lockout + Sign in

+++ Application/Source/InSynq.Core.Service/TokenService.cs
@@ -8,40 +8,40 @@ namespace Nexus.Core.Service;
 
 public class TokenService : ITokenService
 {
-	public string GenerateJwtToken(User user)
-	{
-		var tokenHandler = new JwtSecurityTokenHandler();
+    public string GenerateJwtToken(User user)
+    {
+        var tokenHandler = new JwtSecurityTokenHandler();
 
-		var jwtSecrets = new
-		{
-			Key = Encoding.UTF8.GetBytes(Settings.JwtKey),
-			Settings.Issuer,
-			Settings.Audience,
-		};
+        var jwtSecrets = new
+        {
+            Key = Encoding.UTF8.GetBytes(Settings.JwtKey),
+            Settings.Issuer,
+            Settings.Audience,
+        };
 
-		var claims = new List<Claim>
-		{
-			new(Constants.CLAIM_ID, user.Id.ToString()),
-			new(JwtRegisteredClaimNames.Iss, jwtSecrets.Issuer!),
-			new(Constants.CLAIM_USERNAME, user.Username),
-			new(Constants.CLAIM_EMAIL, user.Email)
-		};
+        var claims = new List<Claim>
+        {
+            new(Constants.CLAIM_ID, user.Id.ToString()),
+            new(JwtRegisteredClaimNames.Iss, jwtSecrets.Issuer!),
+            new(Constants.CLAIM_USERNAME, user.Username),
+            new(Constants.CLAIM_EMAIL, user.Email)
+        };
 
-		var roles = user.Roles.Select(_ => _.RoleId.ToString()).ToList();
+        var roles = user.Roles.Select(_ => ((int)_.RoleId).ToString()).ToList();
 
-		claims.AddRange(roles.Select(role => new Claim(Constants.CLAIM_ROLES, role)));
+        claims.AddRange(roles.Select(role => new Claim(Constants.CLAIM_ROLES, role)));
 
-		var tokenDescriptor = new SecurityTokenDescriptor
-		{
-			Issuer = jwtSecrets.Issuer,
-			Audience = jwtSecrets.Audience,
-			Expires = DateTime.UtcNow.AddDays(Constants.TOKEN_EXPIRATION_TIME),
-			SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(jwtSecrets.Key), SecurityAlgorithms.HmacSha512Signature),
-			Claims = claims.ToDictionary(claim => claim.Type, claim => (object)claim.Value)
-		};
+        var tokenDescriptor = new SecurityTokenDescriptor
+        {
+            Issuer = jwtSecrets.Issuer,
+            Audience = jwtSecrets.Audience,
+            Expires = DateTime.UtcNow.AddDays(Constants.TOKEN_EXPIRATION_TIME),
+            SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(jwtSecrets.Key), SecurityAlgorithms.HmacSha512Signature),
+            Claims = claims.ToDictionary(claim => claim.Type, claim => (object)claim.Value)
+        };
 
-		var token = tokenHandler.CreateToken(tokenDescriptor);
+        var token = tokenHandler.CreateToken(tokenDescriptor);
 
-		return tokenHandler.WriteToken(token);
-	}
+        return tokenHandler.WriteToken(token);
+    }
 }
